Initial commit of dotfiles
These files are used to configure per-user settings for various utilities that
I use on Mac OS X and Linux, and even Windows in some cases.  These files are
pretty raw, and the .vimrc files in particular need a lot of cleanup.

The 'install' script will copy the files out of the working directory into the
appropriate location to be used, typically under $HOME.  It only works for Mac
OS X at this point.

Cleaning out obsolete/unused settings from .[g]vimrc files

Many of these settings were very old and unused for many years.  Some of the
settings were probably ones that I got from someone else but never used.
There are still a few that could probably be cleaned up, particularly related
to the indent options, but this gets rid of most of the cruft.