Adding scripts for installing files on Windows
If the font is already installed, then Windows will pop up a dialog asking
whether it should be reinstalled or not.  It would also be nice to not install
the fonts if they are already installed (or force them to be installed without
prompting), but I don't know of a good way to do that.